add no_update option

+++ bashrc
@@ -1,19 +1,21 @@
 # Are any updates needed?
 ORIGINAL_WD=$(pwd)
 cd ~/.vim
-timeout 2 git remote update > /dev/null 2>&1
-UPDATE_STATUS=$(git status -uno | grep -c -i 'behind\|different')
-if [ "$UPDATE_STATUS" != "0" ]; then
-        read -p "Updates available, update now? " -n 1 -r
-	echo
-        if [[ $REPLY =~ ^[Yy]$ ]]; then
-		git reset --hard
-                git pull -f origin master
-                sh ~/.vim/setup.sh
-                echo "Updates complete."
-        fi
+if [ ! -f '~/.vim/no_update' ]; then
+	timeout 2 git remote update > /dev/null 2>&1
+	UPDATE_STATUS=$(git status -uno | grep -c -i 'behind\|different')
+	if [ "$UPDATE_STATUS" != "0" ]; then
+		read -p "Updates available, update now? " -n 1 -r
+		echo
+		if [[ $REPLY =~ ^[Yy]$ ]]; then
+			git reset --hard
+			git pull -f origin master
+			sh ~/.vim/setup.sh
+			echo "Updates complete."
+		fi
+	fi
+	cd ${ORIGINAL_WD}
 fi
-cd ${ORIGINAL_WD}
 
 TIME='\[\e[96m\]\@'
 WD='\[\e[91m\]\w'
